backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
transformers==4.41.0
Pillow==10.3.0
pydantic==2.7.0
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==2.4.6
numpy==1.26.4
